delay-based-controller: fix invalid pointer copy
When retrieving the connection info again, the copy constructor is
called for all it's fields, including the controller. This means the map
containing the information for all the recorded frames is copied but the
pointers to the most recent two frames are too, and they will still
point to the elements in the old map, rendering them invalid.