scion-capable-node: fix transmission delay computation
Flush transmitted bytes out of the queue before any transmission delay
is computed.

If the queue size is only decreased the moment a packet is entirely
transmitted, the transmission delay that is computed for scheduling
packet sending is wrong. For example, the transmission delay for a new
packet B would be higher than it should be if it's calculated right
before a packet A has been sent, because the queue size will still
include the entire size of A, even though A has already been mostly
transmitted at this point in time.

This even caused packet re-ordering. Assuming two small packets B and C
are queued right after each other. In between, a large packet A leaves
the queue. Now, when B's transmission is scheduled, the queue size is
still large, but when C is scheduled, the queue size will have been
reduced by the size of A. C will then scheduled with a smaller
transmission delay than B and can even arrive at the target before B.

In summary, this change

- prevents false packet re-orders
- improves overall throughput slightly