rtc-app: allow switching paths even without losses
Until now we assumed that, without losses, we haven't yet discovered the
full bottleneck bandwidth of the path and thus shouldn't switch paths
yet. However, if the delay based controller works well and there are no
other flows on the path, there may not be any losses ever. Instead, use
the overuse signal as an indication that we have reached capacity.