tcp: implement wip tcp sender via L3 encapsulation
Implement TCP source and sink applications. This is implemented by
internally creating a dummy IP stack and network, and intercepting
packets at the layer 4 to layer 3 boundary using callback functions.
When a packet is passed from TCP down to the internet layer, we instead
encapsulate it in a SCION packet and deliver it through SCION. At the
other end we again have a TCP application running on a fake IP network,
and feed the decapsulated packet directly to the transport layer
protocol handler. Source and sink both need to be specified and started
separately in the config json file.