simulator: pass app config dict via events json
Main use case is to allow specifying pre-selected paths for simulating
BGP scenarios. General application settings are contained in the json
too.

This is much more flexible and readable than abusing the existing
user-defined events. It would be useful in general to stop using
start_application events and let each application configure itself
entirely using this config map.

Re-using the events.json and making it an extern variable was simply to
save time.